# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/nodes/{id}/content": {
      "get": {
        "operationId": "FilesController_content",
        "parameters": [
          {
            "name": "version",
            "required": false,
            "in": "query",
            "description": "Version id; defaults to the current version",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Location carries the presigned url"
          },
          "410": {
            "description": "Deleted by the owner, or the stored object no longer matches what was verified at upload"
          }
        },
        "security": [
          {
            "share-token": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Redirect to a 5-minute presigned GET for viewing in the browser",
        "tags": [
          "files"
        ]
      }
    },
    "/nodes/{id}/versions": {
      "get": {
        "operationId": "FilesController_list",
        "parameters": [],
        "responses": {
          "404": {
            "description": "Not a file, or no access"
          }
        },
        "security": [
          {
            "share-token": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Version history, newest first",
        "tags": [
          "files"
        ]
      }
    },
    "/nodes/{id}/versions/{versionId}/restore": {
      "post": {
        "operationId": "FilesController_restore",
        "parameters": [
          {
            "name": "versionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Read-only access"
          },
          "422": {
            "description": "That version is already current"
          }
        },
        "security": [
          {
            "share-token": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Make an earlier version current, appended as a new version",
        "tags": [
          "files"
        ]
      }
    },
...
  },
  "info": {
    "title": "Data Room API",
    "description": "Owner-scoped document repository with read-only sharing",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "share-token": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Share-Token"
      }
    },
...
  }
}